Added Body Composition Visualisations
It's worth noting that there is a significant amount of missing data between May 2022 and May 2023. This is likely because I was not strong enough to stand and balance on the scale at that time. I needed to remain still enough for the scales to calculate my body weight (which I couldn't do).

I will note this in the readme