Strength 1RM Data Visualised
When possible, I need to check whether the machines that I was using for legs in 2024 measured weight in lbs or kgs. at the moment the legs charts may be inaccurate.

There may be additional inaccuracies caused by the different machines used. Double pulley machines half the resistance of the weights. Some pulley set ups may not be the same ratio.